// src/app/models/character-model.ts
export class CharacterModel {
    id:                     number = 0;
    nameCharacter:          string;
    roleCharacter:          string;
    descriptionCharacter:   string;
    ageCharacter:           number;
    personalityCharacter:   string;
    skillCharacter:         string;
    createdCharacter:       string;
    nameCreator:            string;
}